+++ optbinning/binning/binning.py
@@ -217,7 +217,7 @@ def _check_parameters(name, dtype, prebinning_method, solver, divergence,
 
     if special_codes is not None:
         if not isinstance(special_codes, (np.ndarray, list, dict)):
-            raise TypeError("special_codes must be a dit, list or "
+            raise TypeError("special_codes must be a dict, list or "
                             "numpy.ndarray.")
 
         if isinstance(special_codes, dict) and not len(special_codes):
@@ -335,7 +335,7 @@ class OptimalBinning(BaseOptimalBinning):
         option currently only applies when ``monotonic_trend`` is "ascending",
         "descending", "peak_heuristic" or "valley_heuristic".
 
-    max_pvalue : float or None, optional (default=0.05)
+    max_pvalue : float or None, optional (default=None)
         The maximum p-value among bins. The Z-test is used to detect bins
         not satisfying the p-value constraint. Option supported by solvers
         "cp" and "mip".