+++ optbinning/binning/multidimensional/binning_2d.py
@@ -551,8 +551,9 @@ class OptimalBinning2D(OptimalBinning):
         return transform_binary_target(
             self.dtype_x, self.dtype_y, self._splits_x_optimal,
             self._splits_y_optimal, x, y, self._n_nonevent, self._n_event,
-            self.special_codes_x, self.special_codes_y, metric, metric_special,
-            metric_missing, show_digits, check_input)
+            self.special_codes_x, self.special_codes_y, self._categories_x,
+            self._categories_y, metric, metric_special, metric_missing,
+            show_digits, check_input)
 
     def _fit(self, x, y, z, check_input):
         time_init = time.perf_counter()
@@ -918,42 +919,6 @@ class OptimalBinning2D(OptimalBinning):
         self._c = c
 
         return rows, n_nonevent, n_event
-
-    def _splits_optimal(self, dtype, selected_rows, splits, P, axis,
-                        categories=None):
-
-        if dtype == "numerical":
-            bins = np.concatenate([[-np.inf], splits, [np.inf]])
-            bins_str = np.array([[bins[i], bins[i+1]]
-                                 for i in range(len(bins) - 1)])
-
-            splits_optimal = []
-            for i in range(len(selected_rows)):
-                if axis == "x":
-                    _, pos = np.where(P == i)
-                else:
-                    pos, _ = np.where(P == i)
-
-                mask = np.arange(pos.min(), pos.max() + 1)
-                bin = bins_str[mask]
-
-                splits_optimal.append([bin[0][0], bin[-1][1]])
-
-            return splits_optimal
-
-        else:
-            splits = np.ceil(splits).astype(int)
-            n_categories = len(categories)
-
-            indices = np.digitize(np.arange(n_categories), splits, right=False)
-            n_bins = len(splits) + 1
-
-            bins = []
-            for i in range(n_bins):
-                mask = (indices == i)
-                bins.append(categories[mask])
-
-            return bins
 
     def _splits_xy_optimal(self, selected_rows, splits_x, splits_y, P):
         bins_x = np.concatenate([[-np.inf], splits_x, [np.inf]])